Classic Restoration Shaman stub
