--- src/analysis/classic/shaman/SPELL_EFFECTS.ts
@@ -1,17 +1,0 @@
-export const EARTH_SHIELD_HEAL = 379;
-export const WATER_SHIELD_ENERGIZE = 23575;
-export const MANA_TIDE_TOTEM_BUFF = 39609;
-export const CLEARCASTING = 16246;
-export const UNLEASHED_RAGE_5 = 30807;
-export const UNLEASHED_RAGE_4 = 30806;
-export const UNLEASHED_RAGE_3 = 30805;
-export const UNLEASHED_RAGE_2 = 30804;
-export const UNLEASHED_RAGE_1 = 30803;
-export const FLURRY_5 = 16280;
-export const FLURRY_4 = 16279;
-export const FLURRY_3 = 16278;
-export const FLURRY_2 = 16277;
-export const FLURRY_1 = 16257;
-export const SHAMANISTIC_FOCUS = 43339;
-export const STORMSTRIKE_MAIN_HAND = 32175;
-export const STORMSTRIKE_OFF_HAND = 32176;
