+++ src/analysis/classic/shaman/restoration/CHANGELOG.tsx
@@ -1,0 +1,6 @@
+import { change, date } from 'common/changelog';
+import { jazminite } from 'CONTRIBUTORS';
+
+export default [
+  change(date(2023, 3, 5), 'Add Classic Restoration Shaman stub.', jazminite),
+];
