+++ src/analysis/classic/shaman/restoration/modules/features/Buffs.tsx
@@ -1,0 +1,27 @@
+import CoreAuras from 'parser/core/modules/Auras';
+import SPELLS from 'common/SPELLS/classic';
+import BLOODLUST_BUFFS from 'game/BLOODLUST_BUFFS';
+import ITEM_BUFFS from 'game/classic/ITEM_BUFFS';
+
+class Buffs extends CoreAuras {
+  // A list of Buffs (on the current player) to highlight on the Timeline
+  auras() {
+    return [
+      {
+        spellId: SPELLS.INNERVATE.id,
+        timelineHighlight: true,
+      },
+      // Do not adjust the lines below
+      {
+        spellId: Object.keys(BLOODLUST_BUFFS).map((item) => Number(item)),
+        timelineHighlight: true,
+      },
+      {
+        spellId: Object.keys(ITEM_BUFFS).map((item) => Number(item)),
+        timelineHighlight: true,
+      },
+    ];
+  }
+}
+
+export default Buffs;
