+++ src/analysis/classic/shaman/restoration/modules/features/CooldownThroughputTracker.ts
@@ -1,0 +1,17 @@
+import SPELLS from 'common/SPELLS/classic';
+import CoreCooldownThroughputTracker, {
+  BUILT_IN_SUMMARY_TYPES,
+} from 'parser/shared/modules/CooldownThroughputTracker';
+
+class CooldownThroughputTracker extends CoreCooldownThroughputTracker {
+  static castCooldowns = [
+    ...CoreCooldownThroughputTracker.castCooldowns,
+    // Add Cooldown Spells specific to Spec
+    {
+      spell: SPELLS.INNERVATE.id,
+      summary: [BUILT_IN_SUMMARY_TYPES.MANA],
+    },
+  ];
+}
+
+export default CooldownThroughputTracker;
