+++ src/analysis/classic/shaman/shared/index.ts
@@ -1,0 +1,2 @@
+export { default as lowRankSpells } from './lowRankSpells';
+export { default as TotemTracker } from './TotemTracker';
