+++ src/analysis/classic/shaman/shared/lowRankSpells.ts
