+++ src/analysis/classic/shaman/shared/totems/FireTotems.tsx
@@ -1,10 +1,10 @@
 import { Options } from 'parser/core/Analyzer';
 
-import { TotemElements } from '../../../totemConstants';
-import TotemTracker from '../../features/TotemTracker';
+import { TotemElements } from './totemConstants';
+import TotemTracker from '../TotemTracker';
 import BaseTotem from './BaseTotem';
 
-export default class AirTotems extends BaseTotem {
+export default class FireTotems extends BaseTotem {
   static dependencies = {
     ...BaseTotem.dependencies,
     totemTracker: TotemTracker,
@@ -13,6 +13,6 @@ export default class AirTotems extends BaseTotem {
   protected totemTracker!: TotemTracker;
 
   constructor(options: Options) {
-    super(options, TotemElements.Air);
+    super(options, TotemElements.Fire);
   }
 }
