+++ src/analysis/classic/shaman/shared/totems/index.ts
@@ -1,0 +1,7 @@
+export { default as AirTotems } from './AirTotems';
+export { default as EarthTotems } from './EarthTotems';
+export { default as FireTotems } from './FireTotems';
+export { default as WaterTotems } from './WaterTotems';
+
+export { default as GroundingTotem } from './GroundingTotem';
+export { default as ManaTideTotem } from './ManaTideTotem';
