+++ src/analysis/classic/shaman/shared/totems/totemConstants.ts
